refactor: extract callToCallable or null to an extra service
