+++ src/language/safe-ds-module.ts
@@ -19,6 +19,7 @@ import { SafeDsValueConverter } from './grammar/safe-ds-value-converter.js';
 import { SafeDsTypeComputer } from './typing/safe-ds-type-computer.js';
 import { SafeDsCoreClasses } from './builtins/safe-ds-core-classes.js';
 import { SafeDsPackageManager } from './workspace/safe-ds-package-manager.js';
+import {SafeDsNodeMapper} from "./helpers/safe-ds-node-mapper.js";
 
 /**
  * Declaration of custom services - add your own service classes here.
@@ -27,6 +28,9 @@ export type SafeDsAddedServices = {
     builtins: {
         CoreClasses: SafeDsCoreClasses;
     };
+    helpers: {
+        NodeMapper: SafeDsNodeMapper;
+    },
     types: {
         TypeComputer: SafeDsTypeComputer;
     };
@@ -49,6 +53,9 @@ export type SafeDsServices = LangiumServices & SafeDsAddedServices;
 export const SafeDsModule: Module<SafeDsServices, PartialLangiumServices & SafeDsAddedServices> = {
     builtins: {
         CoreClasses: (services) => new SafeDsCoreClasses(services),
+    },
+    helpers: {
+        NodeMapper: (services) => new SafeDsNodeMapper(services),
     },
     lsp: {
         Formatter: () => new SafeDsFormatter(),
